Fix incorrect jgit status issue
I've narrowed down the issue with jgit reporting incorrect status of
this repo (repo having changes) while the native 'git status'
returned that the repo was clean.

The culprit seems to be in the SemverReleaseTest tests,
particularly the ones using 'Git.init()' functionality of jgit.

Using 'setGitDir' wasn't correct in the first place, since that
expects a '.git' dir path, not a working directory, but I'm still
not sure how that was causing git status to return incorrect
results.